* apt-pkg/acquire*.{cc,h}:   - try even harder to support really big files in the fetcher by     converting (hopefully) everything to 'long long' (Closes: #632271)

+++ apt-pkg/acquire-method.h
@@ -51,8 +51,8 @@ class pkgAcqMethod
       time_t LastModified;
       bool IMSHit;
       string Filename;
-      unsigned long Size;
-      unsigned long ResumePoint;
+      unsigned long long Size;
+      unsigned long long ResumePoint;
       
       void TakeHashes(Hashes &Hash);
       FetchResult();
